rename HIRT_Multichannel_Convolution/{ConvolveNToMono => n_to_mono}.hpp
